+++ isolate_test.go
@@ -8,15 +8,13 @@ import (
 	"encoding/json"
 	"fmt"
 	"math/rand"
-	"runtime"
 	"strings"
 	"testing"
-	"time"
 
 	v8 "rogchap.com/v8go"
 )
 
-func TestIsolateTermination(t *testing.T) {
+func TestIsolateTerminateExecution(t *testing.T) {
 	t.Parallel()
 	iso := v8.NewIsolate()
 	defer iso.Dispose()
@@ -28,6 +26,9 @@ func TestIsolateTermination(t *testing.T) {
 	var terminating bool
 	fooFn := v8.NewFunctionTemplate(iso, func(info *v8.FunctionCallbackInfo) *v8.Value {
 		loop, _ := info.Args()[0].AsFunction()
+		go func() {
+			iso.TerminateExecution()
+		}()
 		loop.Call(v8.Undefined(iso))
 
 		terminating = iso.IsExecutionTerminating()
@@ -40,12 +41,6 @@ func TestIsolateTermination(t *testing.T) {
 	ctx := v8.NewContext(iso, global)
 	defer ctx.Close()
 
-	go func() {
-		// [RC] find a better way to know when a script has started execution
-		time.Sleep(time.Millisecond)
-		iso.TerminateExecution()
-	}()
-
 	script := `function loop() { while (true) { } }; foo(loop);`
 	_, e := ctx.RunScript(script, "forever.js")
 	if e == nil || !strings.HasPrefix(e.Error(), "ExecutionTerminated") {
@@ -57,7 +52,94 @@ func TestIsolateTermination(t *testing.T) {
 	}
 }
 
-func TestGetHeapStatistics(t *testing.T) {
+func TestIsolateCompileUnboundScript(t *testing.T) {
+	s := "function foo() { return 'bar'; }; foo()"
+
+	i1 := v8.NewIsolate()
+	defer i1.Dispose()
+	c1 := v8.NewContext(i1)
+	defer c1.Close()
+
+	_, err := i1.CompileUnboundScript("invalid js", "filename", v8.CompileOptions{})
+	if err == nil {
+		t.Fatal("expected error")
+	}
+
+	us, err := i1.CompileUnboundScript(s, "script.js", v8.CompileOptions{Mode: v8.CompileModeEager})
+	fatalIf(t, err)
+
+	val, err := us.Run(c1)
+	fatalIf(t, err)
+	if val.String() != "bar" {
+		t.Fatalf("invalid value returned, expected bar got %v", val)
+	}
+
+	cachedData := us.CreateCodeCache()
+
+	i2 := v8.NewIsolate()
+	defer i2.Dispose()
+	c2 := v8.NewContext(i2)
+	defer c2.Close()
+
+	opts := v8.CompileOptions{CachedData: cachedData}
+	usWithCachedData, err := i2.CompileUnboundScript(s, "script.js", opts)
+	fatalIf(t, err)
+	if usWithCachedData == nil {
+		t.Fatal("expected unbound script from cached data not to be nil")
+	}
+	if opts.CachedData.Rejected {
+		t.Fatal("expected cached data to be used, not rejected")
+	}
+
+	val, err = usWithCachedData.Run(c2)
+	fatalIf(t, err)
+	if val.String() != "bar" {
+		t.Fatalf("invalid value returned, expected bar got %v", val)
+	}
+}
+
+func TestIsolateCompileUnboundScript_CachedDataRejected(t *testing.T) {
+	s := "function foo() { return 'bar'; }; foo()"
+	iso := v8.NewIsolate()
+	defer iso.Dispose()
+
+	// Try to compile an unbound script using cached data that does not match this source
+	opts := v8.CompileOptions{CachedData: &v8.CompilerCachedData{Bytes: []byte("Math.sqrt(4)")}}
+	us, err := iso.CompileUnboundScript(s, "script.js", opts)
+	fatalIf(t, err)
+	if !opts.CachedData.Rejected {
+		t.Error("expected cached data to be rejected")
+	}
+
+	ctx := v8.NewContext(iso)
+	defer ctx.Close()
+
+	// Verify that unbound script is still compiled and able to be used
+	val, err := us.Run(ctx)
+	fatalIf(t, err)
+	if val.String() != "bar" {
+		t.Errorf("invalid value returned, expected bar got %v", val)
+	}
+}
+
+func TestIsolateCompileUnboundScript_InvalidOptions(t *testing.T) {
+	iso := v8.NewIsolate()
+	defer iso.Dispose()
+
+	opts := v8.CompileOptions{
+		CachedData: &v8.CompilerCachedData{Bytes: []byte("unused")},
+		Mode:       v8.CompileModeEager,
+	}
+	panicErr := recoverPanic(func() { iso.CompileUnboundScript("console.log(1)", "script.js", opts) })
+	if panicErr == nil {
+		t.Error("expected panic")
+	}
+	if panicErr != "On CompileOptions, Mode and CachedData can't both be set" {
+		t.Errorf("unexpected panic: %v\n", panicErr)
+	}
+}
+
+func TestIsolateGetHeapStatistics(t *testing.T) {
 	t.Parallel()
 	iso := v8.NewIsolate()
 	defer iso.Dispose()
@@ -117,22 +199,60 @@ func TestIsolateDispose(t *testing.T) {
 	}
 }
 
-func TestIsolateGarbageCollection(t *testing.T) {
+func TestIsolateThrowException(t *testing.T) {
 	t.Parallel()
-
 	iso := v8.NewIsolate()
-	val, _ := v8.NewValue(iso, "some string")
-	fmt.Println(val.String())
 
-	tmpl := v8.NewObjectTemplate(iso)
-	tmpl.Set("foo", "bar")
-	v8.NewContext(iso, tmpl)
+	strErr, _ := v8.NewValue(iso, "some type error")
 
+	throwError := func(val *v8.Value) {
+		v := iso.ThrowException(val)
+
+		if !v.IsNullOrUndefined() {
+			t.Error("expected result to be null or undefined")
+		}
+	}
+
+	// Function that throws a simple string error from within the function. It is meant
+	// to emulate when an error is returned within Go.
+	fn := v8.NewFunctionTemplate(iso, func(info *v8.FunctionCallbackInfo) *v8.Value {
+		throwError(strErr)
+
+		return nil
+	})
+
+	// Function that is passed a TypeError from JavaScript.
+	fn2 := v8.NewFunctionTemplate(iso, func(info *v8.FunctionCallbackInfo) *v8.Value {
+		typeErr := info.Args()[0]
+
+		throwError(typeErr)
+
+		return nil
+	})
+
+	global := v8.NewObjectTemplate(iso)
+	global.Set("foo", fn)
+	global.Set("foo2", fn2)
+
+	ctx := v8.NewContext(iso, global)
+
+	_, e := ctx.RunScript("foo()", "foo.js")
+
+	if e.Error() != "some type error" {
+		t.Errorf("expected \"some type error\" error but got: %v", e)
+	}
+
+	_, e = ctx.RunScript("foo2(new TypeError('this is a test'))", "foo.js")
+
+	if e.Error() != "TypeError: this is a test" {
+		t.Errorf("expected \"TypeError: this is a test\" error but got: %v", e)
+	}
+
+	ctx.Close()
 	iso.Dispose()
-
-	runtime.GC()
-
-	time.Sleep(time.Second)
+	if recoverPanic(func() { iso.ThrowException(strErr) }) == nil {
+		t.Error("expected panic")
+	}
 }
 
 func BenchmarkIsolateInitialization(b *testing.B) {