+++ object_test.go
@@ -16,6 +16,8 @@ func TestObjectMethodCall(t *testing.T) {
 
 	ctx := v8.NewContext()
 	iso := ctx.Isolate()
+	defer iso.Dispose()
+	defer ctx.Close()
 	val, _ := ctx.RunScript(`class Obj { constructor(input) { this.input = input, this.prop = "" } print() { return this.input.toString() } }; new Obj("some val")`, "")
 	obj, _ := val.AsObject()
 	val, err := obj.MethodCall("print")
@@ -51,12 +53,24 @@ func TestObjectSet(t *testing.T) {
 	defer ctx.Close()
 	val, _ := ctx.RunScript("const foo = {}; foo", "")
 	obj, _ := val.AsObject()
-	obj.Set("bar", "baz")
+	if err := obj.Set("bar", "baz"); err != nil {
+		t.Errorf("unexpected error: %v", err)
+	}
 	baz, _ := ctx.RunScript("foo.bar", "")
 	if baz.String() != "baz" {
 		t.Errorf("unexpected value: %q", baz)
 	}
-	if err := obj.Set("", nil); err == nil {
+
+	if err := obj.Set("", "zero"); err != nil {
+		t.Errorf("unexpected error: %v", err)
+	}
+	val, err := ctx.RunScript("foo['']", "")
+	fatalIf(t, err)
+	if val.String() != "zero" {
+		t.Errorf("unexpected value: %q", val)
+	}
+
+	if err := obj.Set("a", nil); err == nil {
 		t.Error("expected error but got <nil>")
 	}
 	if err := obj.Set("a", 0); err == nil {
@@ -70,6 +84,53 @@ func TestObjectSet(t *testing.T) {
 	}
 	if ten, _ := ctx.RunScript("foo[10]", ""); ten.String() != "ten" {
 		t.Errorf("unexpected value: %q", ten)
+	}
+}
+
+func TestObjectInternalFields(t *testing.T) {
+	iso := v8.NewIsolate()
+	defer iso.Dispose()
+	ctx := v8.NewContext(iso)
+	defer ctx.Close()
+
+	tmpl := v8.NewObjectTemplate(iso)
+	obj, err := tmpl.NewInstance(ctx)
+	fatalIf(t, err)
+	if count := obj.InternalFieldCount(); count != 0 {
+		t.Errorf("expected 0 got %v", count)
+	}
+	if recoverPanic(func() { obj.GetInternalField(0) }) == nil {
+		t.Error("expected panic")
+	}
+
+	tmpl = v8.NewObjectTemplate(iso)
+	tmpl.SetInternalFieldCount(1)
+	if count := tmpl.InternalFieldCount(); count != 1 {
+		t.Errorf("expected 1 got %v", count)
+	}
+
+	obj, err = tmpl.NewInstance(ctx)
+	fatalIf(t, err)
+	if count := obj.InternalFieldCount(); count != 1 {
+		t.Errorf("expected 1 got %v", count)
+	}
+
+	if v := obj.GetInternalField(0); !v.SameValue(v8.Undefined(iso)) {
+		t.Errorf("unexpected value: %q", v)
+	}
+
+	if err := obj.SetInternalField(0, t); err == nil {
+		t.Error("expected unsupported value error")
+	}
+
+	err = obj.SetInternalField(0, "baz")
+	fatalIf(t, err)
+	if v := obj.GetInternalField(0); v.String() != "baz" {
+		t.Errorf("unexpected value: %q", v)
+	}
+
+	if recoverPanic(func() { obj.SetInternalField(1, "baz") }) == nil {
+		t.Error("expected panic from index out of bounds")
 	}
 }