+++ v8go.h
@@ -6,29 +6,52 @@
 #define V8GO_H
 #ifdef __cplusplus
 
-namespace v8 {
-class Isolate;
-}
+#include "libplatform/libplatform.h"
+#include "v8-profiler.h"
+#include "v8.h"
 
 typedef v8::Isolate* IsolatePtr;
+typedef v8::CpuProfiler* CpuProfilerPtr;
+typedef v8::CpuProfile* CpuProfilePtr;
+typedef const v8::CpuProfileNode* CpuProfileNodePtr;
+typedef v8::ScriptCompiler::CachedData* ScriptCompilerCachedDataPtr;
 
 extern "C" {
 #else
 // Opaque to cgo, but useful to treat it as a pointer to a distinct type
 typedef struct v8Isolate v8Isolate;
 typedef v8Isolate* IsolatePtr;
+
+typedef struct v8CpuProfiler v8CpuProfiler;
+typedef v8CpuProfiler* CpuProfilerPtr;
+
+typedef struct v8CpuProfile v8CpuProfile;
+typedef v8CpuProfile* CpuProfilePtr;
+
+typedef struct v8CpuProfileNode v8CpuProfileNode;
+typedef const v8CpuProfileNode* CpuProfileNodePtr;
+
+typedef struct v8ScriptCompilerCachedData v8ScriptCompilerCachedData;
+typedef const v8ScriptCompilerCachedData* ScriptCompilerCachedDataPtr;
 #endif
 
 #include <stddef.h>
 #include <stdint.h>
 
+// ScriptCompiler::CompileOptions values
+extern const int ScriptCompilerNoCompileOptions;
+extern const int ScriptCompilerConsumeCodeCache;
+extern const int ScriptCompilerEagerCompile;
+
 typedef struct m_ctx m_ctx;
 typedef struct m_value m_value;
 typedef struct m_template m_template;
+typedef struct m_unboundScript m_unboundScript;
 
 typedef m_ctx* ContextPtr;
 typedef m_value* ValuePtr;
 typedef m_template* TemplatePtr;
+typedef m_unboundScript* UnboundScriptPtr;
 
 typedef enum {
   ERROR_RANGE = 1,
@@ -48,12 +71,54 @@ typedef struct {
 } RtnError;
 
 typedef struct {
+  UnboundScriptPtr ptr;
+  int cachedDataRejected;
+  RtnError error;
+} RtnUnboundScript;
+
+typedef struct {
+  ScriptCompilerCachedDataPtr ptr;
+  const uint8_t* data;
+  int length;
+  int rejected;
+} ScriptCompilerCachedData;
+
+typedef struct {
+  ScriptCompilerCachedData cachedData;
+  int compileOption;
+} CompileOptions;
+
+typedef struct {
+  CpuProfilerPtr ptr;
+  IsolatePtr iso;
+} CPUProfiler;
+
+typedef struct CPUProfileNode {
+  CpuProfileNodePtr ptr;
+  const char* scriptResourceName;
+  const char* functionName;
+  int lineNumber;
+  int columnNumber;
+  int childrenCount;
+  struct CPUProfileNode** children;
+} CPUProfileNode;
+
+typedef struct {
+  CpuProfilePtr ptr;
+  const char* title;
+  CPUProfileNode* root;
+  int64_t startTime;
+  int64_t endTime;
+} CPUProfile;
+
+typedef struct {
   ValuePtr value;
   RtnError error;
 } RtnValue;
 
 typedef struct {
-  const char* string;
+  const char* data;
+  int length;
   RtnError error;
 } RtnString;
 
@@ -85,6 +150,26 @@ extern void IsolateTerminateExecution(IsolatePtr ptr);
 extern int IsolateIsExecutionTerminating(IsolatePtr ptr);
 extern IsolateHStatistics IsolationGetHeapStatistics(IsolatePtr ptr);
 
+extern ValuePtr IsolateThrowException(IsolatePtr iso, ValuePtr value);
+
+extern RtnUnboundScript IsolateCompileUnboundScript(IsolatePtr iso_ptr,
+                                                    const char* source,
+                                                    const char* origin,
+                                                    CompileOptions options);
+extern ScriptCompilerCachedData* UnboundScriptCreateCodeCache(
+    IsolatePtr iso_ptr,
+    UnboundScriptPtr us_ptr);
+extern void ScriptCompilerCachedDataDelete(
+    ScriptCompilerCachedData* cached_data);
+extern RtnValue UnboundScriptRun(ContextPtr ctx_ptr, UnboundScriptPtr us_ptr);
+
+extern CPUProfiler* NewCPUProfiler(IsolatePtr iso_ptr);
+extern void CPUProfilerDispose(CPUProfiler* ptr);
+extern void CPUProfilerStartProfiling(CPUProfiler* ptr, const char* title);
+extern CPUProfile* CPUProfilerStopProfiling(CPUProfiler* ptr,
+                                            const char* title);
+extern void CPUProfileDelete(CPUProfile* ptr);
+
 extern ContextPtr NewContext(IsolatePtr iso_ptr,
                              TemplatePtr global_template_ptr,
                              int ref);
@@ -108,6 +193,9 @@ extern void TemplateSetTemplate(TemplatePtr ptr,
 
 extern TemplatePtr NewObjectTemplate(IsolatePtr iso_ptr);
 extern RtnValue ObjectTemplateNewInstance(TemplatePtr ptr, ContextPtr ctx_ptr);
+extern void ObjectTemplateSetInternalFieldCount(TemplatePtr ptr,
+                                                int field_count);
+extern int ObjectTemplateInternalFieldCount(TemplatePtr ptr);
 
 extern TemplatePtr NewFunctionTemplate(IsolatePtr iso_ptr, int callback_ref);
 extern RtnValue FunctionTemplateGetFunction(TemplatePtr ptr,
@@ -117,7 +205,7 @@ extern ValuePtr NewValueNull(IsolatePtr iso_ptr);
 extern ValuePtr NewValueUndefined(IsolatePtr iso_ptr);
 extern ValuePtr NewValueInteger(IsolatePtr iso_ptr, int32_t v);
 extern ValuePtr NewValueIntegerFromUnsigned(IsolatePtr iso_ptr, uint32_t v);
-extern RtnValue NewValueString(IsolatePtr iso_ptr, const char* v);
+extern RtnValue NewValueString(IsolatePtr iso_ptr, const char* v, int v_length);
 extern ValuePtr NewValueBoolean(IsolatePtr iso_ptr, int v);
 extern ValuePtr NewValueNumber(IsolatePtr iso_ptr, double v);
 extern ValuePtr NewValueBigInt(IsolatePtr iso_ptr, int64_t v);
@@ -127,7 +215,7 @@ extern RtnValue NewValueBigIntFromWords(IsolatePtr iso_ptr,
                                         int word_count,
                                         const uint64_t* words);
 extern ValuePtr NewValueError(IsolatePtr iso_ptr, ErrorTypeIndex idx, const char* message);
-const char* ValueToString(ValuePtr ptr);
+extern RtnString ValueToString(ValuePtr ptr);
 const uint32_t* ValueToArrayIndex(ValuePtr ptr);
 int ValueToBoolean(ValuePtr ptr);
 int32_t ValueToInt32(ValuePtr ptr);
@@ -197,8 +285,11 @@ const char* ExceptionGetMessageString(ValuePtr ptr);
 
 extern void ObjectSet(ValuePtr ptr, const char* key, ValuePtr val_ptr);
 extern void ObjectSetIdx(ValuePtr ptr, uint32_t idx, ValuePtr val_ptr);
+extern int ObjectSetInternalField(ValuePtr ptr, int idx, ValuePtr val_ptr);
+extern int ObjectInternalFieldCount(ValuePtr ptr);
 extern RtnValue ObjectGet(ValuePtr ptr, const char* key);
 extern RtnValue ObjectGetIdx(ValuePtr ptr, uint32_t idx);
+extern ValuePtr ObjectGetInternalField(ValuePtr ptr, int idx);
 int ObjectHas(ValuePtr ptr, const char* key);
 int ObjectHasIdx(ValuePtr ptr, uint32_t idx);
 int ObjectDelete(ValuePtr ptr, const char* key);